#오류수정

+++ main.py
@@ -6,7 +6,7 @@ from idlelib import paragraph
 from random import random, randint
 from strategy import full_para_
 
-from pip.cmdoptions import cache_dir
+#from pip.cmdoptions import cache_dir
 
 import os
 from manager import db_manager, log_manager
@@ -95,7 +95,7 @@ if __name__ == '__main__':
         print('#%d.\t\t 테스트 변수 (strategy.json, test.json) 읽어오는 중 ...' % step.__next__())
 
     strategy_var = json_reader.Reader.read_strategy_config()
-    start_date, end_date = json_reader.Reader.read_test_config()
+    start_date, end_date, viewer = json_reader.Reader.read_test_config()
     real_start_date = "20990101"
     real_end_date = "20000101"
 
@@ -139,11 +139,15 @@ if __name__ == '__main__':
 
 
     for file_name in cache_dir:
-        file_path = '%s/%s/%s' % (os.path.curdir, '/cached_candles', file_name)
-        file_datetime = datetime.strptime(time.ctime(os.path.getctime(file_path)), "%a %b %d %H:%M:%S %Y")
-        if file_datetime.day is not datetime.now().day:
-            os.remove(file_path)
-            print('Remove Cached File(%s)' % file_name)
+        if file_name == "GCM19_tick_60":
+            file_path = '%s/%s/%s' % (os.path.curdir, '/cached_candles', file_name)
+            file_datetime = datetime.strptime(time.ctime(os.path.getctime(file_path)), "%a %b %d %H:%M:%S %Y")
+            if file_datetime.day is not datetime.now().day:
+                os.remove(file_path)
+                print('Remove Cached File(%s)' % file_name)
+                cache_dir = os.listdir('%s/%s' % (os.path.curdir, '/cached_candles'))
+        else:
+            pass
 
     for chart in strategy_var[CHARTS]:
         if main_chart is None:
@@ -179,6 +183,10 @@ if __name__ == '__main__':
 
     tmp_candles = {}
     # chart_candles 변환
+
+    backend_data = {}
+
+
     for chart_id in chart_candles.keys():
         tmp_candles[chart_id] = {}
         tmp_candles[chart_id][시가] = []
@@ -190,6 +198,9 @@ if __name__ == '__main__':
         tmp_candles[chart_id][영업일] = []
         tmp_candles[chart_id][가격들] = []
 
+        backend_data[chart_id] = []
+
+        index = 0
         for candle in chart_candles[chart_id]:
             if not start_date <= candle[영업일] <= end_date:
                 continue
@@ -204,7 +215,9 @@ if __name__ == '__main__':
             tmp_candles[chart_id][체결시간].append(datetime.strptime(candle[체결시간], '%Y-%m-%d %H:%M:%S'))
             tmp_candles[chart_id][거래량].append(candle[거래량])
             tmp_candles[chart_id][영업일].append(candle[영업일])
-            tmp_candles[chart_id][가격들].append([float(price) for price in candle[가격들].split(',')])
+            prices = [float(price) for price in candle[가격들].split(',')]
+            tmp_candles[chart_id][가격들].append(prices)
+
 
     '''상단까지가 우리가 입력한 날짜에 맞는 테이블을 Tick_60 으로만 가져오는 코드'''
     start_time = time.time()
@@ -222,6 +235,8 @@ if __name__ == '__main__':
 
         if TEST_MAIN_LOG:
             print('#%d.\t\t 병렬 테스트 수행 (Core 수=%d, 횟수=%d)' % (step.__next__(), (mp.cpu_count() - 1), total_count))
+            if total_count > 1:
+                viewer_data = None
 
         pool = mp.Pool(processes=mp.cpu_count() - 1)
         #pool = mp.Pool(processes=mp.cpu_count() * 2)
@@ -231,7 +246,7 @@ if __name__ == '__main__':
             config = StrategyVarManager.get_speific_startegy_var(cur_array)
             #pprint(config)
             ''' 해당 부분에서 Multiprocessing 으로 테스트 시작 '''
-            procs_results.append(pool.apply_async(func=simulator.simulate, args=(main_chart, config, common_candles,),
+            procs_results.append(pool.apply_async(func=simulator.simulate, args=(main_chart, config, common_candles, viewer, ),
                                                   callback=end_simulate))
 
             if StrategyVarManager.increase_the_number_of_digits(max_array, cur_array) is False:
@@ -268,8 +283,8 @@ if __name__ == '__main__':
                 except ZeroDivisionError:
                     win_lose = 0
 
-                print('\t\t %s: %s   승률:%s,   승:%s, 패:%s' % (report.종목코드, report.수익, win_lose, report.승, report.패 ))
-                fprint('\t\t %s: %s   승률:%s,   승:%s, 패:%s' % (report.종목코드, report.수익, win_lose, report.승, report.패))
+                print('\t\t %s: %s   승률:%s,   승:%s, 패:%s, 최대수익:%s, 최대손실:-%s' % (report.종목코드, report.수익, win_lose, report.승, report.패, report.최대수익, report.최대손실 ))
+                fprint('\t\t %s: %s   승률:%s,   승:%s, 패:%s, 최대수익:%s, 최대손실:-%s' % (report.종목코드, report.수익, win_lose, report.승, report.패, report.최대수익, report.최대손실))
 
     #full_para_.FullPara.calc_reports(simulation_report)
 
@@ -282,6 +297,6 @@ if __name__ == '__main__':
     print("DB에 저장!!!!!")
     print('test_start_date=%s \t real_start_date=%s' % (start_date, real_start_date))
     print('test_end_date=%s \t real_end_date=%s' % (end_date, real_end_date))
-    dbm.insert_test_result(simulation_report[0], real_start_date, real_end_date)
+    #dbm.insert_test_result(simulation_report[0], real_start_date, real_end_date)
